Update some unit tests
