+++ lib/commands/touch.js
@@ -259,7 +259,7 @@ function getCoordDefault(val) {
  * @param {number} digits
  * @returns {number}
  */
-function truncateDecimals(number, digits) {
+export function truncateDecimals(number, digits) {
   const multiplier = Math.pow(10, digits),
     adjustedNum = number * multiplier,
     truncatedNum = Math[adjustedNum < 0 ? 'ceil' : 'floor'](adjustedNum);
